+++ lib/src/utils.dart
@@ -10,23 +10,23 @@ import 'package:dart_style/dart_style.dart';
 import 'package:path/path.dart';
 
 const List keywords = const [
-  "assert", "break", "case", "catch", "class", "const", "continue",
-  "default", "do", "else", "enum", "extends", "false", "final", "finally",
-  "for", "if", "in", "is", "new", "null", "rethrow", "return", "super",
-  "switch",
-  "this", "throw", "true", "try", "var", "void", "while", "with",
+  'assert', 'break', 'case', 'catch', 'class', 'const', 'continue',
+  'default', 'do', 'else', 'enum', 'extends', 'false', 'final', 'finally',
+  'for', 'if', 'in', 'is', 'new', 'null', 'rethrow', 'return', 'super',
+  'switch',
+  'this', 'throw', 'true', 'try', 'var', 'void', 'while', 'with',
 
   // This is not in the dart language specification 1.2 but is reserved
   // in dart2js and the dart VM.
   // See: http://dartbug.com/19515
-  "external"
+  'external'
 ];
 
-final _cleanRegEx = new RegExp(r"[^\w$]");
+final _cleanRegEx = new RegExp(r'[^\w$]');
 
 String fileDate(DateTime date) =>
     "${date.year}${(date.month < 10) ? 0 : ""}${date.month}${(date.day < 10) ? 0 : ""}${date.day}_${(date.hour < 10) ? 0 : ""}${date.hour}${(date.minute < 10) ? 0 : ""}${date.minute}${(date.second < 10) ? 0 : ""}${date.second}";
-String cleanName(String name) => name.replaceAll(_cleanRegEx, "_");
+String cleanName(String name) => name.replaceAll(_cleanRegEx, '_');
 
 final DartFormatter formatter =
     new DartFormatter(lineEnding: '\n', pageWidth: 80);
@@ -90,10 +90,10 @@ String findPackageRoot(String path) {
   return null;
 }
 
-const String gitIgnore = """
+const String gitIgnore = '''
 packages
 pubspec.lock
-""";
+''';
 
 class GenerateResult {
   final String apiName;
@@ -128,7 +128,7 @@ class GenerateResult {
   }
 
   String get shortName =>
-      cleanName("${apiName}_${apiVersion}_api").toLowerCase();
+      cleanName('${apiName}_${apiVersion}_api').toLowerCase();
 
   String toString() {
     if (info) {